Create etch-a-sketch container
Add a div container in index.html where all the boxes will be stored.
In javascript, use css to flex this container, set a max width
and flex-wrap so at maximum, it will be a 16x16 grid. Then create a
for loop that creates a new div inside the container, with a black
solid border and specific height and width.