Add comments, reword variables, add random colors
Add a function that returns a random color hexcode so that on
a hover, the color is random rather than being one set color.
Add comments and reword some variables so it is easier to read. Add
a couple more styles to overall container such as flex and shorten
the overall height and width of grid.